+++ src/db/dexie/repositories/index.ts
@@ -10,3 +10,4 @@ export * from "./DexieManualRepository";
 export * from "./DexieProjectsRepository";
 export * from "./DexieSettingsRepository";
 export * from "./DexieTasksRepository";
+export * from "./DexieRoutinesRepository";